Create CartRowRemoveButton to reuse between coupons and items
